benchmark: prototype

+++ src/SandViewer/Scene_load.cpp
@@ -159,7 +159,7 @@ bool Scene::load(const std::string & filename)
 
 	const glm::vec2 & res = viewportCamera()->resolution();
 	DEBUG_LOG << "reso: " << res.x << ", " << res.y;
-	m_deferredShader.setResolution(res.x, res.y);
+	m_deferredShader.setResolution(static_cast<int>(res.x), static_cast<int>(res.y));
 
 	reloadShaders();
